refactor: 구조 리팩토링 (#416)
* refactor: 이전 PR에서 반영되지 않은 구조 반영 (#415)

* refactor: VideoQuery 패키지에 dao 적용 (#415)

* refactor: Auth 패키지에 dao 적용 (#415)

* refactor: Restaurant 패키지에 dao 적용 (#415)

* style: polishing (#415)

---------
