+++ backend/src/main/java/com/celuveat/admin/exception/AdminExceptionType.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpStatus;
 public enum AdminExceptionType implements BaseExceptionType {
 
     NOT_EXISTS_CELEB(BAD_REQUEST, "셀럽이 저장되어 있지 않습니다. 셀럽 먼저 저장해주세요."),
-    ILLEGAL_DATE_FORMAT(BAD_REQUEST, "영상 업로드 날짜 형식이 잘못됐습니다. \'yyyy. M. d\' 형식으로 입력해주세요."),
+    ILLEGAL_DATE_FORMAT(BAD_REQUEST, "영상 업로드 날짜 형식이 잘못됐습니다. 'yyyy. M. d' 형식으로 입력해주세요."),
     ILLEGAL_FORMAT(BAD_REQUEST, "데이터 저장 양식이 잘못됐습니다. 엑셀 폼을 준수해주세요."),
     EXIST_NULL(BAD_REQUEST, "입력되지 않은 값이 있습니다."),
     INVALID_YOUTUBE_CHANNEL_NAME(BAD_REQUEST, "유튜브 채널 명이 잘못됐습니다. 앞에 '@'를 붙여주세요."),
