+++ backend/src/main/java/com/celuveat/auth/query/dao/OauthMemberQueryDaoSupport.java
@@ -1,4 +1,4 @@
-package com.celuveat.auth.query;
+package com.celuveat.auth.query.dao;
 
 import static com.celuveat.auth.exception.AuthExceptionType.NOT_FOUND_MEMBER;
 
@@ -6,7 +6,7 @@ import com.celuveat.auth.command.domain.OauthMember;
 import com.celuveat.auth.exception.AuthException;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OauthMemberQueryRepository extends JpaRepository<OauthMember, Long> {
+public interface OauthMemberQueryDaoSupport extends JpaRepository<OauthMember, Long> {
 
     default OauthMember getById(Long id) {
         return findById(id).orElseThrow(() -> new AuthException(NOT_FOUND_MEMBER));
