+++ backend/src/main/java/com/celuveat/restaurant/query/dao/RestaurantImageQueryDaoSupport.java
@@ -1,11 +1,11 @@
-package com.celuveat.restaurant.query;
+package com.celuveat.restaurant.query.dao;
 
 import com.celuveat.restaurant.command.domain.Restaurant;
 import com.celuveat.restaurant.command.domain.RestaurantImage;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantImageQueryRepository extends JpaRepository<RestaurantImage, Long> {
+public interface RestaurantImageQueryDaoSupport extends JpaRepository<RestaurantImage, Long> {
 
     List<RestaurantImage> findAllByRestaurantIdIn(List<Long> restaurantIds);
