+++ backend/src/main/java/com/celuveat/restaurant/query/dao/RestaurantQueryDaoSupport.java
@@ -1,4 +1,4 @@
-package com.celuveat.restaurant.query;
+package com.celuveat.restaurant.query.dao;
 
 import static com.celuveat.restaurant.exception.RestaurantExceptionType.NOT_FOUND_RESTAURANT;
 
@@ -6,8 +6,7 @@ import com.celuveat.restaurant.command.domain.Restaurant;
 import com.celuveat.restaurant.exception.RestaurantException;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantQueryRepository extends
-        JpaRepository<Restaurant, Long>, RestaurantEntityManagerQueryRepository {
+public interface RestaurantQueryDaoSupport extends JpaRepository<Restaurant, Long> {
 
     @Override
     default Restaurant getById(Long id) {
