+++ backend/src/test/java/com/celuveat/restaurant/fixture/LocationFixture.java
@@ -1,6 +1,6 @@
 package com.celuveat.restaurant.fixture;
 
-import com.celuveat.restaurant.query.RestaurantEntityManagerQueryRepositoryImpl.LocationSearchCond;
+import com.celuveat.restaurant.query.dao.RestaurantWithDistanceDao.LocationSearchCond;
 import com.celuveat.restaurant.query.dto.RestaurantSimpleResponse;
 
 public class LocationFixture {
